+++ Sources/ABAlbum/Views/AlbumSelectorGridCellView.swift
@@ -6,15 +6,36 @@
 //
 
 import SwiftUI
+import Photos
 
 struct AlbumSelectorGridCellView: View {
-    var image: Image?
+    
+    var asset: PHAsset
+    var size: CGSize
+    var thumbnailSize: CGSize
+    var requestOptions: PHImageRequestOptions?
     var title: String?
     var count: Int?
     
+    @State private var thumbnailImage: UIImage?
+    
+    @ViewBuilder
+    var image: some View {
+        if thumbnailImage == nil {
+            Image("photo_demo", bundle: .module)
+        } else {
+            Image(uiImage: thumbnailImage!)
+        }
+    }
+    
     var body: some View {
-        VStack(alignment: .leading) {
-            Spacer()
+        ZStack(alignment: .bottom) {
+            Image(uiImage: thumbnailImage ?? UIImage())
+                .resizable()
+                .scaledToFill()
+                .frame(width: size.width, height: size.height)
+                .clipped()
+                .overlay(ProgressView().frame(width: size.width, height: size.height).background(Color(uiColor: .tertiarySystemGroupedBackground)).opacity(thumbnailImage == nil ? 1 : 0))
             HStack {
                 Text(title ?? "Untitled")
                 Spacer()
@@ -23,21 +44,30 @@ struct AlbumSelectorGridCellView: View {
             .padding(10)
             .background(.thinMaterial)
         }
-        .background(
-            image ?? Image("photo_demo", bundle: .module)
-        )
-        .clipped()
+        .onAppear {
+            guard thumbnailImage == nil else { return }
+            Task(priority: .high) {
+                do {
+                    for try await image in AlbumService.shared.asyncImage(from: asset, size: thumbnailSize, requestOptions: requestOptions) {
+                        thumbnailImage = image
+                    }
+                } catch let error {
+                    thumbnailImage = nil
+                    print(error)
+                }
+            }
+        }
     }
 }
 
 struct AlbumSelectorGridCellView_Previews: PreviewProvider {
     static var previews: some View {
         Group {
-            AlbumSelectorGridCellView()
+            AlbumSelectorGridCellView(asset: PHAsset(), size: CGSize.zero, thumbnailSize: CGSize.zero)
                 .preferredColorScheme(.dark)
                 .previewLayout(.sizeThatFits)
                 .frame(width: 200.0, height: 200.0)
-            AlbumSelectorGridCellView()
+            AlbumSelectorGridCellView(asset: PHAsset(), size: CGSize.zero, thumbnailSize: CGSize.zero)
                 .preferredColorScheme(.light)
                 .previewLayout(.sizeThatFits)
                 .frame(width: 200.0, height: 200.0)